backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
stripe==11.5.0

twilio==9.3.0
